Add rotation and a few related helper functions
This was pulled out of Vault's codebase, then enhanced to split
different parts of it into different functions. This also introduces the
option pattern. The rotation test validates creation and deletion
independently as well as the combination via the rotate function.

Coverage (when running with real AWS creds) is mostly green for all
added code, except for places where we are handling exceptional
conditions from AWS' API.