awsutil: add GetCallerIdentity to Create/Rotate (#10)
* awsutil: add GetCallerIdentity to Create/Rotate

This adds GetCallerIdentity to CreateAccessKey and RotateKeys. The
latter is done simply by passing WithTimeout through to the inner
CreateAccessKey on Rotate. As such, both paths are conditional on a
non-zero WithTimeout option being given - this is to ensure that we
don't try to immediately verify, which is likely to always fail.

* awsutil: add some comments, remove exponential backoff

This does the following:

* Adds some comments for WithTimeout to RotateKeys and CreateAccessKey
  to explain exactly what these options do.

* Removes the exponential backoff. Even with the quadratic-time backoff
  we're still probably looking at a number of seconds of delay that's
  probably unnecessary. Retrying every 1s should not generate that much
  traffic and only need to be necessary for a few retries.

* Change WithTimeout to WithValidityCheckTimeout