awsutil: add ability to mock IAM and STS APIs
This adds the ability to supply mock IAM and STS interfaces to use with
the RotateKeys, CreateAccessKey, DeleteAccessKey, and GetCallerIdentity
methods.

Additionally, work has been done to incorporate the existing IAM mock
object into the package, in addition to adding a similar STS mock that
allows for the mocking of the GetCallerIdentity method.

Factory functions are also included to allow these to be incorporated
seamlessly into the call path, in addition to allowing for introspection
into any session data being received by the constructor functions.