feature: Add support handling forwarded for headers
This PR adds TrustedFromXForwardedFor(...) which uses the
XForwardedFor* listener cfg setting to determine how/if headers are
trusted/allowed for inbound http requests.

It also adds, WrapForwardedForHandler(...) which is an http middleware
wrapper that uses TrustedFromXForwardedFor to determine how/if
headers are trusted/allowed for inbound http requests

This handler wrapper is lifted from vault v1.8.5 with the intention
of sharing it between both boundary and vault.