Propagate plugin options in ConfigureWrapper (#41)
* fix: Make testplugin functional

The testplugin as previously implemented didn't
support the Encrypt or Decrypt methods,
since it did not have an AEAD cipher configured.
This change configures an AES cipher with
a hardcoded key, so that we can do Encrypt
and Decrypt operations in tests.

* fix: Propagate plugin options in configureWrapper

The configureWrapper was not propagating
plugin options. Propagate these options
and add a test to ensure they work.