listenerutil: Support http.Flusher, http.Hijacker, http.Pusher
Using the new http.ResponseController in Go 1.20, we
can easily add support for http.Flusher, http.Hijacker, http.Pusher
and the SetReadDeadline and SetWriteDeadline methods to
the ResponseWriter wrapper. This fixes some services that
rely on this interface being implemented for their functionality.

In Go 1.19, just implement Flusher, Pusher and Hijacker explicitly.