feat(awsutil): major version update
This major version release utilizes the latest version
of the aws-sdk-go-v2. The following behavioral changes
are included in this major version release:

- Custom endpoint resolvers are attached to the STS and
IAM clients, not to the credentials. This is apart of the
aws-sdk-go-v2 EndpointResolverV2 feature.
- withStsEndpoint is no longer a string type, but a
sts.EndpointResolverV2 type. This option was relabeled
to withStsEndpointResolver.
- withIamEndpoint is no longer a string type, but a
iam.EndpointResolverV2 type. This option was relabeled
to withIamEndpointResolver.
- By default, aws credential configurations will load values
from environment variables. The user provided options will
overload the default values.
- The ability to mock out the underlying credential provider
for unit testing.

Changed behaviors from awsutil v1 includes the following:

- Replaced aws errors with aws smithy-go errors
- No longer able to utilize the aws default remote credential
provider
- The function GenerateCredentialChain returns a aws.Config,
which contains the credential provider.