Fix client.Start() race condition, more debug/trace logs
Set Follow: true for container logs, without which there was an inadvertent race condition
that normally succeeded on my laptop but failed on CI. The plugin had to have already printed
its protocol negotiation line on stdout by the time we started reading logs or else the
logging function would get to the end of the logs at that time and return, causing go-plugin
to pick up an empty string and say 'negotiation failed'. This was a pig to track down.

Also adds some extra debug/trace logs to shed light on the plugin lifecycle.

Finally, disable cgo when building the plugin for tests, so that it successfully runs in the
static distroless image where libc is not available.