Fix double-enter issue on Windows (#95)
Currently, running Read twice on Windows will result in the second call immediately exiting. This is because the first call leaves a \n in the stdin buffer on Windows since Windows sends \r\n on Enter.

This change causes us to loop around again on Windows when we receive an \r so that we can then consume the \n and drain it. This fixes the issue. Note the caveats in the comment, which I'm not worried about.